chore: ProductPrice 이 int 가 아닌 BigDecimal 을 값으로 갖도록 수정

+++ src/main/java/kitchenpos/products/tobe/domain/model/vo/ProductPrice.java
@@ -1,14 +1,15 @@
 package kitchenpos.products.tobe.domain.model.vo;
 
+import java.math.BigDecimal;
 import kitchenpos.global.domain.vo.ValueObject;
 import org.springframework.util.Assert;
 
 public class ProductPrice extends ValueObject {
 
-    private final int price;
+    private final BigDecimal price;
 
     public ProductPrice(int price) {
         Assert.isTrue(price >= 0, "상품 가격은 0원 미만일 수 없습니다.");
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
     }
 }
